simauth: fix gsmAuthenticate
The input to the comp128 algorithm was relying on QSTRING_TO_BUF
returning a pointer to persistant memory. This was not the case
as it was returning a pointer to an intermediate object which was
being freed once out of scope. It just happened to work most of
the time. This change copies ki/rand into a static buffer.